requirements.txt
python-telegram-bot==11.1.0
dnspython==1.16.0
requests==2.7.0
pymongo==3.8.0
google==2.0.2
google-api-python-client==1.7.8
urllib3==1.24.1
oauth2client==4.1.3
matplotlib==3.1.3
turtle==0.0.2

